refactor: make `Switch` component re-usable in example app
