+++ FabricExample/src/components/Switch/index.tsx
