+++ FabricExample/src/screens/Examples/AwareScrollView/index.tsx
@@ -3,10 +3,10 @@ import React, { useCallback, useEffect, useRef, useState } from "react";
 import { Button, Platform, Text, View } from "react-native";
 import { KeyboardAwareScrollView } from "react-native-keyboard-controller";
 
+import Switch from "../../../components/Switch";
 import TextInput from "../../../components/TextInput";
 
 import { styles } from "./styles";
-import Switch from "./Switch";
 
 import type { ExamplesStackParamList } from "../../../navigation/ExamplesStack";
 import type { StackScreenProps } from "@react-navigation/stack";
