+++ example/src/components/Switch/index.tsx
